+++ tests/e2e/examples/bobsbooks/bobs_books_test.go
@@ -75,8 +75,8 @@ var afterSuite = t.AfterSuiteFunc(func() {
 		dump.CaptureContainerLogs(namespace, "bobbys-front-end-adminserver", "weblogic-server", "/scratch/logs/bobbys-front-end")
 		dump.CaptureContainerLogs(namespace, "bobbys-front-end-managed-server1", "weblogic-server", "/scratch/logs/bobbys-front-end")
 		// Bobs Bookstore
-		dump.CaptureContainerLogs(namespace, "bobs-bookstore-adminserver", "weblogic-server", "/scratch/logs/bobs-orders-wls")
-		dump.CaptureContainerLogs(namespace, "bobs-bookstore-managed-server1", "weblogic-server", "/scratch/logs/bobs-orders-wls")
+		dump.CaptureContainerLogs(namespace, "bobs-bookstore-adminserver", "weblogic-server", "/scratch/logs/bobs-bookstore")
+		dump.CaptureContainerLogs(namespace, "bobs-bookstore-managed-server1", "weblogic-server", "/scratch/logs/bobs-bookstore")
 		dump.ExecuteBugReport(namespace)
 	}
 	if !skipUndeploy {
@@ -342,7 +342,7 @@ var _ = t.Describe("Bobs Books test", Label("f:app-lcm.oam",
 				},
 				func() {
 					Eventually(func() bool {
-						return pkg.MetricsExist("istio_tcp_received_bytes_total", "destination_canonical_service", "bobs-orders-wls")
+						return pkg.MetricsExist("istio_tcp_received_bytes_total", "destination_canonical_service", "bobs-bookstore")
 					}, longWaitTimeout, longPollingInterval).Should(BeTrue())
 				},
 				func() {